split up expression and context into separate files
